requirements.txt
python-telegram-bot==13.2
flask==2.2.2
werkzeug==2.2.3
python-dotenv==0.20.0
hurry.filesize==0.9
pathvalidate==3.2.0
